chore: clear the clippy backlog
Six warnings that had accumulated: two redundant closures, a
filter().next_back() that rfind says in one call, an owned String built
only to be compared, a helper stranded below its own test module, and a
doc comment that had drifted away from the function it describes — it now
sits on the memoised keychain read it was written for.

No behaviour change; the memo comment is inside the release-only module,
so debug builds never compiled it either way.

+++ src-tauri/tests/workflow.rs
@@ -595,7 +595,7 @@ fn large_libraries_page_in_sql_and_stay_responsive() {
             file_size: Some(1024 * 1024),
             width: Some(3840),
             height: Some(2160),
-            duration_ms: Some((i % 120) as i64 * 1000),
+            duration_ms: Some((i % 120) * 1000),
             source_created_at: None,
             source_modified_at: None,
             tags: None,
